Moved into ydb/core/util
